[~] Fix pull request

+++ api/IGraphicLib.hpp
@@ -25,7 +25,7 @@ namespace Arcade {
 		// Main loop condition
 		virtual bool isOpen() const = 0;
 
-		// Closes the window => stop loop (go menu)
+		// Closes the window => stop loop (go start)
 		virtual void closeRenderer() = 0;
 
 		// Opens the window => start loop
@@ -66,9 +66,9 @@ namespace Arcade {
 		virtual Vect<size_t> getScreenSize() const = 0;
 
 		// get the Y max of the windows
-		virtual int getMaxY() const = 0;
+		virtual size_t getMaxY() const = 0;
 
 		// get the X max of the windows
-		virtual int getMaxX() const = 0;
+		virtual size_t getMaxX() const = 0;
 	};
 };
